Add IDFAFetcher as part of SDK
